This is probably one of the last commits on the project TalkMyPhone itself. This commit makes TalkMyPhone a dummy application that connects to jabber and broadcasts the messages it receives on the system. It also listens to some broadcasted intents (to be documented soon) and transmits the inner messages to the user.
The logic is moved to another project named android rules that I just
started:

hg clone https://androidrules.googlecode.com/hg/ androidrules

androidrules is an app that can talk to TalkMyPhone. The aim is to allow
the user to control his phone through simple rules.
I intent to make an open-source implementation of Tasker out of it.

For now, the androidrules project is in an alpha-state (it does not really
serve any purpose without TalkMyPhone, the code is still tied to what
TalkMyPhone does). The aim of androidrules is to become a real project
that can run without TalkMyPhone.